# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Examination & OSM Platform",
    "description": "Backend API for AI-assisted Digital Examination & On-Screen Marking Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/students": {
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Create Student Endpoint",
        "description": "S1: Add a new candidate.",
        "operationId": "create_student_endpoint_api_v1_students_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Search Students Endpoint",
        "description": "S2: Search candidates (paginated).",
        "operationId": "search_students_endpoint_api_v1_students_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term for roll number or full name",
              "title": "Q"
            },
            "description": "Search term for roll number or full name"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{id}/answer-sheets": {
      "post": {
        "tags": [
          "Answer Sheets, Mapping & Assignments"
        ],
        "summary": "Upload Answer Sheet Endpoint",
        "description": "AS1: Upload a student's answer sheet (PDF or JPEG/PNG pages).",
        "operationId": "upload_answer_sheet_endpoint_api_v1_exams__id__answer_sheets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_answer_sheet_endpoint_api_v1_exams__id__answer_sheets_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerSheetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Answer Sheets, Mapping & Assignments"
        ],
        "summary": "List Answer Sheets Endpoint",
        "description": "AS2: List answer sheets with progress (Admin only).",
        "operationId": "list_answer_sheets_endpoint_api_v1_exams__id__answer_sheets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerSheetListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/answer-sheets/{id}": {
      "get": {
        "tags": [
          "Answer Sheets, Mapping & Assignments"
        ],
        "summary": "Get Answer Sheet Detail Endpoint",
        "description": "AS3: Get answer sheet detail, pages, and answers (role-aware blind marking).",
        "operationId": "get_answer_sheet_detail_endpoint_api_v1_answer_sheets__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerSheetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pages/{page_id}/image": {
      "get": {
        "tags": [
          "Answer Sheets, Mapping & Assignments"
        ],
        "summary": "Stream Page Image Endpoint",
        "description": "AS5: Stream a page image (authorized for admin, assigned examiner, or moderator).",
        "operationId": "stream_page_image_endpoint_api_v1_pages__page_id__image_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{id}/assignments": {
      "post": {
        "tags": [
          "Answer Sheets, Mapping & Assignments"
        ],
        "summary": "Assign Examiners Endpoint",
        "description": "AS6: Assign unmarked answers to examiners (by_sheet or by_question).",
        "operationId": "assign_examiners_endpoint_api_v1_exams__id__assignments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerSheetDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exam_id": {
            "type": "integer",
            "title": "Exam Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "anon_code": {
            "type": "string",
            "title": "Anon Code"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploaded",
              "mapped"
            ],
            "title": "Status"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "student": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "mapped": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapped"
          },
          "answers_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers Total"
          },
          "answers_marked": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers Marked"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/AnswerSheetPageResponse"
            },
            "type": "array",
            "title": "Pages",
            "default": []
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerSummary"
            },
            "type": "array",
            "title": "Answers",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "exam_id",
          "anon_code",
          "status",
          "page_count",
          "created_at"
        ],
        "title": "AnswerSheetDetailResponse"
      },
      "AnswerSheetListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnswerSheetResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "AnswerSheetListResponse"
      },
      "AnswerSheetPageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "sheet_id": {
            "type": "integer",
            "title": "Sheet Id"
          },
          "page_number": {
            "type": "integer",
            "title": "Page Number"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sheet_id",
          "page_number",
          "image_url"
        ],
        "title": "AnswerSheetPageResponse"
      },
      "AnswerSheetResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "exam_id": {
            "type": "integer",
            "title": "Exam Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "anon_code": {
            "type": "string",
            "title": "Anon Code"
          },
          "status": {
            "type": "string",
            "enum": [
              "uploaded",
              "mapped"
            ],
            "title": "Status"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "student": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "mapped": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapped"
          },
          "answers_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers Total"
          },
          "answers_marked": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answers Marked"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exam_id",
          "anon_code",
          "status",
          "page_count",
          "created_at"
        ],
        "title": "AnswerSheetResponse"
      },
      "AnswerSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "question_number": {
            "type": "string",
            "title": "Question Number"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "is_attempted": {
            "type": "boolean",
            "title": "Is Attempted"
          },
          "ocr_status": {
            "type": "string",
            "title": "Ocr Status"
          },
          "ai_status": {
            "type": "string",
            "title": "Ai Status"
          },
          "marking_status": {
            "type": "string",
            "title": "Marking Status"
          },
          "final_marks": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Marks"
          },
          "max_marks": {
            "type": "number",
            "title": "Max Marks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "question_id",
          "question_number",
          "is_attempted",
          "ocr_status",
          "ai_status",
          "marking_status",
          "max_marks"
        ],
        "title": "AnswerSummary"
      },
...
      "AssignmentRequest": {
        "properties": {
          "examiner_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Examiner Ids"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "by_sheet",
              "by_question"
            ],
            "title": "Strategy",
            "default": "by_sheet"
          }
        },
        "type": "object",
        "required": [
          "examiner_ids"
        ],
        "title": "AssignmentRequest"
      },
      "AssignmentResponse": {
        "properties": {
          "assigned": {
            "type": "integer",
            "title": "Assigned"
          },
          "per_examiner": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Per Examiner"
          }
        },
        "type": "object",
        "required": [
          "assigned",
          "per_examiner"
        ],
        "title": "AssignmentResponse"
      },
...
      "Body_upload_answer_sheet_endpoint_api_v1_exams__id__answer_sheets_post": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "files"
        ],
        "title": "Body_upload_answer_sheet_endpoint_api_v1_exams__id__answer_sheets_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StudentCreate": {
        "properties": {
          "roll_number": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Roll Number"
          },
          "full_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Full Name"
          },
          "department": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          }
        },
        "type": "object",
        "required": [
          "roll_number",
          "full_name"
        ],
        "title": "StudentCreate"
      },
      "StudentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "StudentListResponse"
      },
      "StudentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "roll_number": {
            "type": "string",
            "title": "Roll Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "roll_number",
          "full_name",
          "created_at"
        ],
        "title": "StudentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}